+++ src/index.js
@@ -1,5 +1,5 @@
 import './style.scss';
-import { addToDom , cryptoCount, getData} from './modules/Home';
+import { addToDom, cryptoCount, getData } from './modules/Home';
 import { request } from './modules/api';
 
 const coins = document.querySelector('.coins');
